LayoutTests/platform/chromium-win/inspector/debugger-autocontinue-on-syntax-error-expected.txt

CONSOLE MESSAGE: line 1: Uncaught SyntaxError: Unexpected token )
CONSOLE MESSAGE: line 1: Uncaught SyntaxError: Unexpected token )
Tests that debugger won't stop on syntax errors even if "pause on uncaught exceptions" is on.

Debugger was enabled.
Page reloaded.
DONE.
